add user_service and rollback oauth service
