+++ registry/services/oauth/token_service.py
@@ -1,11 +1,10 @@
 from typing import Optional, List, Dict, Any, Tuple
-from beanie import PydanticObjectId
 from packages.models._generated.token import Token
-from packages.models._generated.user import IUser
 from registry.models.oauth_models import OAuthTokens
 from registry.models.emus import TokenType
 from registry.utils.log import logger
 from datetime import datetime, timezone, timedelta
+from registry.services.user_service import user_service
 
 
 class TokenService:
@@ -38,7 +37,7 @@ class TokenService:
             Created or updated Token document
         """
         identifier = self._get_client_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         # Calculate expiration time
         expires_at = self._calculate_expiration(tokens.expires_in)
@@ -95,7 +94,7 @@ class TokenService:
             return None
 
         identifier = self._get_refresh_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         # Refresh tokens typically have a longer expiration time, set to 1 year here
         # Or set according to OAuth provider configuration
@@ -194,14 +193,14 @@ class TokenService:
             Token document or None
         """
         identifier = self._get_client_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         token = await Token.find_one({
             "userId": user_obj_id,
             "type": TokenType.MCP_OAUTH_CLIENT.value,
             "identifier": identifier
         })
-        logger.debug(f"OAuth client token for user={user_id}, token={token}")
+        logger.debug(f"OAuth client token for user={user_id}, service={service_name}")
 
         # Check if token is expired
         if token and self._is_token_expired(token):
@@ -226,7 +225,7 @@ class TokenService:
             Token document or None
         """
         identifier = self._get_refresh_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         token = await Token.find_one({
             "userId": user_obj_id,
@@ -287,7 +286,7 @@ class TokenService:
         Returns:
             Whether deletion was successful
         """
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         # Delete client token
         client_identifier = self._get_client_identifier(service_name)
@@ -335,7 +334,7 @@ class TokenService:
         Returns:
             List of Token documents
         """
-        user_obj_id = await self._get_user_object_id(user_id)
+        user_obj_id = await user_service.get_user_object_id(user_id)
 
         query = {"userId": user_obj_id}
         if token_type:
@@ -421,17 +420,17 @@ class TokenService:
                 - is_valid: True if token exists and not expired, False otherwise
         """
         identifier = self._get_client_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
-        
+        user_obj_id = await user_service.get_user_object_id(user_id)
+
         token = await Token.find_one({
             "userId": user_obj_id,
             "type": TokenType.MCP_OAUTH_CLIENT.value,
             "identifier": identifier
         })
-        
+
         if not token:
             return None, False
-        
+
         is_valid = not self._is_token_expired(token)
         return token, is_valid
 
@@ -449,79 +448,19 @@ class TokenService:
                 - is_valid: True if token exists and not expired, False otherwise
         """
         identifier = self._get_refresh_identifier(service_name)
-        user_obj_id = await self._get_user_object_id(user_id)
-        
+        user_obj_id = await user_service.get_user_object_id(user_id)
+
         token = await Token.find_one({
             "userId": user_obj_id,
             "type": TokenType.MCP_OAUTH_REFRESH.value,
             "identifier": identifier
         })
-        
+
         if not token:
             return None, False
-        
+
         is_valid = not self._is_token_expired(token)
         return token, is_valid
-
-    async def _get_user_object_id(self, user_id: str) -> PydanticObjectId:
-        """
-        Get or create user's ObjectId  TODO: Will be replaced with real user system
-        
-        Args:
-            user_id: User ID (username)
-            
-        Returns:
-            User's PydanticObjectId
-        """
-        user = await IUser.find_one({"username": user_id})
-
-        if not user:
-            now = datetime.now(timezone.utc)
-            email = f"{user_id}@local.mcp-gateway.internal"
-
-            existing_user = await IUser.find_one({"email": email})
-            if existing_user:
-                return existing_user.id
-
-            # Create user
-            user_data = {
-                "username": user_id,
-                "email": email,
-                "emailVerified": False,
-                "role": "USER",
-                "provider": "local",
-                "createdAt": now,
-                "updatedAt": now
-            }
-
-            collection = IUser.get_pymongo_collection()
-            result = await collection.insert_one(user_data)
-            logger.info(f"Created user record for token storage: {user_id}")
-            return result.inserted_id
-
-        return user.id
-
-    async def get_or_create_user(self, email: str) -> Optional[IUser]:
-        """
-        Get user by email
-        """
-        user = await IUser.find_one({"email": email})
-        if not user:
-            now = datetime.now(timezone.utc)
-            # Create user
-            user_data = {
-                "email": email,
-                "emailVerified": False,
-                "role": "USER",
-                "provider": "local",
-                "createdAt": now,
-                "updatedAt": now
-            }
-            collection = IUser.get_pymongo_collection()
-            await collection.insert_one(user_data)
-            logger.info(f"Created user record for token storage: {email}")
-            user = await IUser.find_one({"email": email})
-        return user
 
     def _calculate_expiration(self, expires_in: Optional[int]) -> datetime:
         """
